chore: move api/ into runtime/
